+++ Finandroid/app/src/main/java/com/finandroid/models/Transaction.kt
@@ -1,13 +1,13 @@
 package com.finandroid.models
 
 class Transaction {
-    val value: Double;
+    val value: Float;
     val bank: String;
     val category: String;
     val description: String;
     val flow: Int;
 
-    constructor(value: Double, bank: String, category: String, description: String, flow: Int) {
+    constructor(value: Float, bank: String, category: String, description: String, flow: Int) {
         this.value = value
         this.bank = bank
         this.category = category
@@ -18,7 +18,7 @@ class Transaction {
 
 class TransactionResponse {
     val id: Long;
-    val value: Double;
+    val value: Float;
     val bank: String;
     val category: String;
     val description: String;
@@ -26,7 +26,7 @@ class TransactionResponse {
 
     constructor(
         id: Long,
-        value: Double,
+        value: Float,
         bank: String,
         category: String,
         description: String,
